feat(world): Phase 4 全版本兼容 — ChunkPayloadParser SPI + 1.13-1.17/1.12 legacy 适配
- ChunkPayloadParser SPI：AnvilWorldReader 按区块 NBT 形状逐块分派
  （现代 sections / 1.13-1.17 Level+Palette / 1.12- 数字 ID），
  升级存档混合格式天然支持，解析器注册表可注入扩展
- PalettedContainer 支持 1.13-1.15 旧跨 long 打包（crossLong），
  新增 fromIndices 重建容器（修复 int 位移溢出）
- NumericLegacyChunkParser：Blocks/Add/Data nibble 经程序化
  flattening 映射表（LegacyFlatteningTable）转为扁平化状态，
  未知 (id,meta) 降级 stone 并计入 unknownStats 覆盖统计
- PalettedLegacyChunkParser：Level 包装、DataVersion<2529 旧打包、
  256 列 / 1024 quart 群系重排为 section 级 4x4x4 容器
- LegacyBiomeIds：1.12 群系 ID → 现代等价染色名（近似映射）
- testFixtures 新增 LegacySyntheticWorldBuilder（1.12 数值 /
  1.15 调色板两种布局，含跨 long 打包器）
- golden 测试：flattening 映射/Add nibble/nibble 光照/列群系、
  跨 long 调色板/quart 群系按全局 Y 抽取、混合存档逐块分派回归
- 修复 LegacyFlatteningTable FACING6 越界（meta 6/7 钳位）
- README 路线图 Phase 4 标记完成

+++ modules/world-context/src/testFixtures/java/online/yudream/voxelith/world/testfixtures/SyntheticWorldBuilder.java
@@ -180,7 +180,7 @@ public final class SyntheticWorldBuilder {
         for (int i = 0; i < indices.length; i++) {
             int longIndex = i / perLong;
             int slot = i % perLong;
-            data[longIndex] |= (indices[i] & mask) << (slot * bits);
+            data[longIndex] |= ((long) indices[i] & mask) << (slot * bits);
         }
         return data;
     }
